crt: string: updated the return type of `strlen()'.
The `strlen()' function now returns an integer of type `size_t' instead
of `int'. This is dictated by POSIX.

+++ include/etaos/string.h
@@ -24,7 +24,7 @@
 CDECL
 
 extern char *strchr(const char *str, int c);
-extern int strlen(const char *str);
+extern size_t strlen(const char *str);
 extern int strnlen(const char *str, size_t size);
 extern int strcmp(const char *s1, const char *s2);
 extern void *memset(void *dst, int c, size_t n);
